Store polygon indices as u16, not u64. Maps from 743MB total down to 620MB!

+++ widgetry/src/backend_glow.rs
@@ -217,7 +217,7 @@ impl PrerenderInnards {
         let mut indices: Vec<u32> = Vec::new();
 
         for (color, poly) in batch.consume() {
-            let idx_offset = vertices.len();
+            let idx_offset = vertices.len() as u32;
             let (pts, raw_indices) = poly.raw_for_rendering();
             for pt in pts {
                 let style = color.style(*pt);
@@ -231,7 +231,7 @@ impl PrerenderInnards {
                 ]);
             }
             for idx in raw_indices {
-                indices.push((idx_offset + *idx) as u32);
+                indices.push(idx_offset + (*idx as u32));
             }
         }
